Remove circular references prohibited by default since spring boot 2.6
