+++ src/main/java/io/insee/dev/k8sonboarding/controller/ClusterAccessController.java
@@ -1,10 +1,7 @@
 package io.insee.dev.k8sonboarding.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -12,7 +9,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import io.insee.dev.k8sonboarding.configuration.security.UserProvider;
-import io.insee.dev.k8sonboarding.model.User;
 import io.insee.dev.k8sonboarding.service.OnboardingService;
 import io.insee.dev.k8sonboarding.view.ClusterCredentials;
 
@@ -26,12 +22,6 @@ public class ClusterAccessController {
     @Autowired
     private UserProvider userProvider;
 
-    @Value("${io.insee.dev.k8sonboarding.jwt.username-claim}")
-    private String usernameClaim;
-
-    @Value("${io.insee.dev.k8sonboarding.jwt.groups-claim:groups}")
-    private String groupsClaim;
-
     @GetMapping
     public ClusterCredentials getCredentials(Authentication auth) throws IllegalAccessException {
         return getCredentials(auth, null);
@@ -39,7 +29,7 @@ public class ClusterAccessController {
 
     @GetMapping("/credentials/{groupId}")
     public ClusterCredentials getCredentials(Authentication auth, @PathVariable String groupId) throws IllegalAccessException {
-	return onboardingService.getClusterCredentials(userProvider.getUser(auth), groupId);
+        return onboardingService.getClusterCredentials(userProvider.getUser(auth), groupId);
     }
 
     @PostMapping("/namespace")
@@ -49,29 +39,17 @@ public class ClusterAccessController {
 
     @PostMapping("/namespace/{groupId}")
     public void createNamespaceForGroup(Authentication auth, @PathVariable String groupId) {
-	onboardingService.createNamespace(userProvider.getUser(auth), groupId);
+        onboardingService.createNamespace(userProvider.getUser(auth), groupId);
     }
 
     @PostMapping("/permissions")
     public void addPermissionsForUser(Authentication auth) {
-	onboardingService.addPermissionsToNamespace(userProvider.getUser(auth), null);
+        onboardingService.addPermissionsToNamespace(userProvider.getUser(auth), null);
     }
 
     @PostMapping("/permissions/{groupId}")
     public void addPermissionsForUser(Authentication auth, @PathVariable String groupId) {
         onboardingService.addPermissionsToNamespace(userProvider.getUser(auth), groupId);
-    }
-
-    @Bean
-    public UserProvider getUserProvider() {
-	return auth -> {
-	    final User user = new User();
-	    final Jwt jwt = (Jwt) auth.getPrincipal();
-	    user.setId(jwt.getClaimAsString(usernameClaim));
-	    user.setGroups(jwt.getClaimAsStringList(groupsClaim));
-	    user.setAuthToken(jwt.getTokenValue());
-	    return user;
-	};
     }
 
 }
